Instruction decoders cleanup part 2: instruction tracing:
Rename GENx370x390x900 macro arguments to make clearer that the "_format" argument is NOT the format of the instruction itself (as used by the instruction decoders), but instead identifies the PRINT format that should be used during instruction tracing, which has NOTHING to do with the instruction's actual machine code format.

I eventually want to rename our instruction decoder macros to match the same names as used in the Principles of Operation (without impacting instruction tracing/printing logic), but need to first eliminate the confusion caused by the poorly named GENx370x390x900 "format" argument, which simply routes instruction TRACING to a given print function but otherwise has NOTHING to do with the instruction DECODER format (which is quite misleading and confusing!).