Stop panel select watching logger pipe:
Remove unnecessary reference to the logger pipe file descriptor in the main panel loop. This caused the select() to continually pop if data was present in the pipe because this thread did not drain the data. This manifested itself on uniprocessor systems when running as root as a CPU burn in the panel thread.

Closes #289 Panel issue under Linux.